add : Bugs 봇 구현
벅스 크롤링 봇 구현
